+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## NEW FEATURES
 
 - require('@openveo/api').util.getPropertyFromArray can now start collecting values at a specified value
+- require('@openveo/api').storages.ResourceFilter now supports regular expressions
 
 # 6.0.0 / 2018-10-16