+++ CHANGELOG.md
@@ -6,6 +6,17 @@
 - require('@openveo/api').storages.ResourceFilter now supports regular expressions
 - require('@openveo/api').util.escapeTextForRegExp has been added to help escape a text that will be used in a JavaScript regular expression
 
+## DEPENDENCIES
+
+- **chai** has been upgraded from 4.0.2 to **4.2.0**
+- **chai-spies** has been upgraded from 0.7.1 to **1.0.0**
+- **grunt** has been upgraded from 1.0.1 to **1.0.3**
+- **grunt-eslint** has been upgraded from 19.0.0 to **21.0.0**
+- **grunt-gh-pages** has been upgraded from 2.0.0 to **3.1.0**
+- **grunt-mocha-test** has been upgraded from 0.13.2 to **0.13.3**
+- **mocha** has been upgraded from 3.2.0 to **5.2.0**
+- **mock-require** has been upgraded from 3.0.1 to **3.0.2**
+
 # 6.0.0 / 2018-10-16
 
 ## BREAKING CHANGES